requirements.txt
openai>=0.27.0
gradio>=3.0
pillow>=9.0.0
python-dotenv>=0.21.0
requests>=2.28.0
